requirements.txt
streamlit
qrcode
Pillow
io